requirements.txt
flask
streamlit
ultralytics
pillow
matplotlib
gdown
torch